Use the comment_as_submitted array for comparing the current comment to the last-seen comment.
The real comment data can be modified/sanitized before it reaches matches_last_comment(), resulting in false negatives -- comments placed in moderation that Akismet has marked as spam.

I've seen it in practice when the email address uses Chinese characters -- something strips them out of the comment passed to matches_last_comment() but not out of the comment we stored in Akismet::$last_comment.


git-svn-id: http://plugins.svn.wordpress.org/akismet/trunk@884214 b8457f37-d9ea-0310-8a92-e5e31aec5664