Prevent floods of moderation emails.
If Akismet ever goes down (or if it's unreachable from a given server for a reason other than downtime), it's undesirable to flood the user with moderation emails for comments that Akismet will probably mark as spam as soon as it's available again.

This changeset delays moderation emails for 24 hours; if Akismet continues to be unavailable, moderation emails will be sent out during the first Akismet retry after 24 hours.

If Akismet becomes available again before 24 hours, moderation emails will only be sent for comments that are not spam and that merit a moderation email.


git-svn-id: http://plugins.svn.wordpress.org/akismet/trunk@886566 b8457f37-d9ea-0310-8a92-e5e31aec5664